pyproject.toml
[tool.poetry]
name = "k8s-image-updater"
version = "0.1.0"
description = ""
authors = ["Andrin Kessler <andrin.kessler@ik.me>"]
readme = "README.md"
packages = [{include = "k8s_image_updater"}]

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
